change private to public + move directory
