+++ src/main/java/fr/openmc/core/features/events/contents/dailyevents/contents/goldenharvest/GoldenHarvestEvent.java
@@ -9,8 +9,8 @@ import fr.openmc.core.features.events.models.HasMenu;
 import fr.openmc.core.lifecycle.interfaces.HasFeature;
 import fr.openmc.core.registry.ambient.CustomAmbient;
 import fr.openmc.core.registry.features.Feature;
-import fr.openmc.core.registry.features.FeatureEntry;
-import fr.openmc.core.registry.features.FeatureLoadingType;
+import fr.openmc.core.registry.features.loading.FeatureEntry;
+import fr.openmc.core.registry.features.loading.FeatureLoadingType;
 import fr.openmc.core.utils.nms.toast.CustomToastData;
 import fr.openmc.core.utils.text.messages.TranslationManager;
 import net.kyori.adventure.bossbar.BossBar;
