+++ src/main/java/fr/openmc/core/lifecycle/interfaces/HasFeature.java
@@ -1,7 +1,7 @@
 package fr.openmc.core.lifecycle.interfaces;
 
 import fr.openmc.core.registry.features.Feature;
-import fr.openmc.core.registry.features.FeatureEntry;
+import fr.openmc.core.registry.features.loading.FeatureEntry;
 
 /**
  * Interface permettant aux classes d'enregistrer une feature
