+++ src/main/java/fr/openmc/core/registry/features/loading/FeatureEntry.java
@@ -1,8 +1,9 @@
-package fr.openmc.core.registry.features;
+package fr.openmc.core.registry.features.loading;
 
 import com.google.common.base.Supplier;
 import fr.openmc.core.OMCPlugin;
 import fr.openmc.core.lifecycle.integration.OMCLogger;
+import fr.openmc.core.registry.features.Feature;
 import lombok.Getter;
 
 import java.util.Set;
